Move grid cell component into its own file
