+++ packages/react/src/Grid/Grid.tsx
@@ -12,7 +12,7 @@ import {
   PropsWithChildren,
   RefAttributes,
 } from 'react'
-import { gridCellClasses } from './gridCellClasses'
+import { GridCell } from './GridCell'
 
 export type GridColumnNumber = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12
 
@@ -99,41 +99,5 @@ export const Grid = forwardRef(
   ),
 ) as GridComponent
 
-type GridCellFullWidthProp = {
-  /** Whether the cell spans the full width of the grid. */
-  fullWidth?: boolean
-  span?: never
-  start?: never
-}
-
-type GridCellColumnProps = {
-  fullWidth?: never
-  /** The amount of grid columns the cell spans. */
-  span?: GridColumnNumber | GridColumnNumbers
-  /** The index of the grid column the cell starts at. */
-  start?: GridColumnNumber | GridColumnNumbers
-}
-
-export type GridCellProps = (GridCellFullWidthProp | GridCellColumnProps) &
-  PropsWithChildren<HTMLAttributes<HTMLDivElement>>
-
-const GridCell = forwardRef(
-  ({ children, className, fullWidth, span, start, ...restProps }: GridCellProps, ref: ForwardedRef<HTMLDivElement>) => (
-    <div
-      {...restProps}
-      ref={ref}
-      className={clsx(
-        'amsterdam-grid__cell',
-        fullWidth && 'amsterdam-grid__cell--full-width',
-        gridCellClasses(span, start),
-        className,
-      )}
-    >
-      {children}
-    </div>
-  ),
-)
-
 Grid.Cell = GridCell
 Grid.displayName = 'Grid'
-GridCell.displayName = 'Grid.Cell'
