+++ packages/react/src/Grid/index.ts
@@ -1,2 +1,2 @@
 export * from './Grid'
-export type { GridCellProps } from './Grid'
+export type { GridCellProps } from './GridCell'
